+++ src/test/java/contactapp/service/ProjectServiceLegacyTest.java
@@ -35,13 +35,12 @@ class ProjectServiceLegacyTest extends PostgresContainerSupport {
     @Autowired
     private ProjectService springService;
 
+    @Autowired
+    private TestCleanupUtility testCleanup;
+
     @BeforeEach
     void resetSingleton() throws Exception {
-        org.springframework.security.core.context.SecurityContextHolder.clearContext();
-        testUserSetup.cleanup();
-        testUserSetup.setupTestUser();
-        springService.clearAllProjects();
-        setInstance(null);
+        testCleanup.resetTestEnvironment();
     }
 
     @AfterEach